Move testQueryLoggingCount to AbstractTestEngineOnlyQueries
